NASABASE/requirements.txt
chromadb
openai
tqdm
numpy
python-dotenv
fastapi
uvicorn
chromadb
openai
python-dotenv